+++ build/build-eucalyptus-console-selinux-rpm.sh
@@ -2,6 +2,7 @@
 # Build eucalyptus console selinux rpm on CentOS/RHEL 7
 
 # config
+MODE="${1:-build}" # setup build build-only
 EUCA_CON_SE_BRANCH="master"
 EUCA_CON_SE_REPO="https://github.com/eucalyptus/eucaconsole-selinux.git"
 REQUIRE=(
@@ -14,20 +15,26 @@ REQUIRE=(
     "yum-utils"
 )
 set -ex
-RPMBUILD=${RPMBUILD:-$(mktemp -td "rpmbuild.XXXXXXXXXX")}
 
 # dependencies
-yum erase -y 'eucaconsole-*'
+if [ "${MODE}" != "build-only" ] ; then
+  yum erase -y 'eucaconsole-*'
 
-yum -y install "${REQUIRE[@]}"
+  yum -y install "${REQUIRE[@]}"
 
-yum -y groupinstall development
+  yum -y groupinstall development
+fi
+
+[ "${MODE}" != "setup" ] || exit 0
 
 # clone repositories
-[ ! -d "eucaconsole-selinux" ] || rm -rf "eucaconsole-selinux"
-git clone --depth 1 --branch "${EUCA_CON_SE_BRANCH}" "${EUCA_CON_SE_REPO}"
+if [ "${MODE}" != "build-only" ] ; then
+  [ ! -d "eucaconsole-selinux" ] || rm -rf "eucaconsole-selinux"
+  git clone --depth 1 --branch "${EUCA_CON_SE_BRANCH}" "${EUCA_CON_SE_REPO}"
+fi
 
 # setup rpmbuild
+RPMBUILD=${RPMBUILD:-$(mktemp -td "rpmbuild.XXXXXXXXXX")}
 mkdir -p "${RPMBUILD}/SPECS"
 mkdir -p "${RPMBUILD}/SOURCES"
 
@@ -48,11 +55,15 @@ tar -cvJf "${RPMBUILD}/SOURCES/eucaconsole-selinux-${EUCA_CON_SE_VERSION}.tar.xz
     "eucaconsole-selinux"
 
 # build rpms
+RPMBUILD_OPTS="${RPMBUILD_OPTS}"
+RPM_DIST="${RPM_DIST:-el7}"
 RPM_VERSION="$(date -u +%Y%m%d)git"
+RPM_BUILD_ID="${RPM_BUILD_ID:-${RPM_VERSION}${EUCA_CON_SE_GIT_SHORT}}"
 
 rpmbuild \
     --define "_topdir ${RPMBUILD}" \
-    --define "dist .${RPM_VERSION}${EUCA_CON_SE_GIT_SHORT}.el7" \
+    --define "dist .${RPM_BUILD_ID}.${RPM_DIST}" \
+    ${RPMBUILD_OPTS} \
     -ba "${RPMBUILD}/SPECS/eucaconsole-selinux.spec"
 
 find "${RPMBUILD}/SRPMS/"
